Move touchpad event types into their own module
